requirements.txt
fastapi
uvicorn
langchain
openai
huggingface_hub
pandas
yfinance
python-dotenv
